Prep dynamic workspace loading
